+++ src/mesido/pycml/component_library/milp/multicommodity/__init__.py
@@ -1,11 +1,11 @@
 from .airwater_heat_pump_elec import AirWaterHeatPumpElec
 from .elec_heat_source_elec import ElecHeatSourceElec
 from .electrolyzer import Electrolyzer
-from .gas_boiler import GasBoiler
+from .gas_heat_source_gas import GasHeatSourceGas
 
 __all__ = [
     "AirWaterHeatPumpElec",
     "ElecHeatSourceElec",
     "Electrolyzer",
-    "GasBoiler",
+    "GasHeatSourceGas",
 ]