+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Cost attributes of GasHeatSourceGas (gas boiler with 3 ports) are supported in Mesido
 - HeatSourceElec (e-boiler with 2 ports) is supported for DTK and its cost attributes are supported in Mesido
 - Cost attributes of ElecHeatSourceElec (e-boiler with 3 ports) are supported in Mesido
+- Maximum profile constraint for PV asset is considered in PV sizing
 
 ## Changed
 - Clean up of old code and removing duplicates.